Adding Resnet50 width variation and moving resnet functions to separate file instead of backbone.py (Issue #34)
